refactor: move schema validation to a graphql mod
