+++ src/blueprint/from_config/graphql/mod.rs
@@ -1,0 +1,3 @@
+mod schema;
+
+pub use schema::*;
