+++ src/blueprint/from_config/mod.rs
@@ -1,5 +1,9 @@
 mod from_config;
 mod server;
+mod graphql;
+mod types;
 
 pub use server::*;
 pub use from_config::*;
+pub use graphql::*;
+pub use types::*;
