+++ src/blueprint/from_config/types.rs
@@ -1,0 +1,3 @@
+use crate::{try_fold::TryFold, config::Config};
+
+pub type TryFoldConfig<'a, A> = TryFold<'a, Config, A, String>;
